implementing podbox-specific ballot configs
