+++ apps/zupoll-client/src/types.ts
@@ -1,4 +1,4 @@
-import { BallotType, LoginConfig } from "@pcd/zupoll-shared";
+import { LoginConfig } from "@pcd/zupoll-shared";
 
 export type ZupollError = {
   /** Big title, should be under 40 chars */
@@ -13,18 +13,6 @@ export enum PCDState {
   DEFAULT,
   AWAITING_PCDSTR,
   RECEIVED_PCDSTR
-}
-
-export interface BallotConfig {
-  voterGroupId: string;
-  voterGroupUrl: string;
-  creatorGroupId: string;
-  creatorGroupUrl: string;
-  passportServerUrl: string;
-  passportAppUrl: string;
-  ballotType: BallotType;
-  latestGroupHashUrl?: string;
-  makeHistoricalGroupUrl?: (hash: string) => string;
 }
 
 export interface LoginState {
